# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: feedspool HTTP API
  version: "1"
  description: |
    Read and annotate a feedspool database over HTTP.

    Enable with `feedspool serve --api`. Authentication is off unless a token
    is configured (`serve.api.token` or `FEEDSPOOL_API_TOKEN`); when one is
    set, every endpoint requires `Authorization: Bearer <token>`.

    Compatibility: within v1, changes are additive only -- new fields, new
    optional parameters, new endpoints, new error codes, new `include` values.
    Removing or renaming a field, changing a type, or changing a parameter's
    default requires v2. The cursor encoding is explicitly outside the
    contract: treat it as an opaque token.

    Unknown query parameters are rejected with 400 rather than ignored, so a
    typo surfaces immediately instead of silently returning a default page.
servers:
  - url: /api/v1
    description: Mounted on the serve command
security:
  - bearerAuth: []
  - {}
paths:
...
  /api/v1/items:
    get:
      summary: List items
      description: |
        Keyset-paginated, newest first by default. Offset paging is not offered
        because items arrive at the head of the ordering continuously, which
        makes offsets skip and duplicate rows mid-scan.

        Note that `archived` defaults to `false` here, which differs from
        `feedspool items`, where archived rows are included.
      operationId: listItems
      parameters:
        - name: feed_id
          in: query
          description: Filter to one feed by derived ID. Mutually exclusive with `feed_url` and `feed_query`.
          schema: { type: string }
        - name: feed_url
          in: query
          description: Filter to one feed by exact URL. Mutually exclusive with `feed_id` and `feed_query`.
          schema: { type: string }
        - name: feed_query
          in: query
          description: Filter to feeds whose URL contains this substring, case-insensitively.
          schema: { type: string }
        - name: link
          in: query
          description: Exact item link. May match several items across feeds.
          schema: { type: string }
        - name: q
          in: query
          description: |
            Case-insensitive substring match on the item title only -- not body
            or summary. Identical to `feedspool items --search`.
          schema: { type: string }
        - name: since
          in: query
          description: |
            RFC3339. Filters on discovery time (`first_seen`, falling back to
            `published_date`), not publication date, because feeds routinely
            backdate `published_date`.
          schema: { type: string, format: date-time }
        - name: until
          in: query
          description: RFC3339, on discovery time.
          schema: { type: string, format: date-time }
        - name: seen
          in: query
          description: Filter by presence of a `seen` annotation. Omit for both.
          schema: { type: string, enum: ["true", "false"] }
        - name: archived
          in: query
          description: Defaults to `false`. Use `any` for both.
          schema: { type: string, enum: ["true", "false", "any"], default: "false" }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, oldest], default: newest }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
        - name: include
          in: query
          description: Comma-separated. One of `content`, `raw`, `metadata`, `annotations`.
          schema: { type: string }
      responses:
        "200":
          description: A page of items
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Collection"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Item" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Collection:
      type: object
      required: [data, next_cursor, limit]
      description: |
        The envelope every list endpoint returns. There is deliberately no
        total: it costs a second full scan and is stale by the time a client
        reads it.
      properties:
        data:
          type: array
          items: {}
        next_cursor:
          type: [string, "null"]
          description: Opaque token for the next page, or null when exhausted.
        limit:
          type: integer
          description: The effective page size, after any clamping.
    Item:
      type: object
      required:
        [id, feed_id, feed_url, guid, title, link, summary, published_date,
         first_seen, discovered_at, archived]
      properties:
        id: { type: string }
        feed_id: { type: string }
        feed_url: { type: string }
        guid: { type: string }
        title: { type: string }
        link: { type: string }
        summary: { type: string }
        published_date:
          type: [string, "null"]
          format: date-time
          description: Null for scraped feeds, which deliberately leave it unset.
        first_seen:
          type: [string, "null"]
          format: date-time
        discovered_at:
          type: [string, "null"]
          format: date-time
          description: The sort key -- published_date falling back to first_seen.
        archived: { type: boolean }
        content:
          type: string
          description: Present only with include=content.
        item_json:
          description: Raw parsed source. Present only with include=raw.
        annotations:
          type: array
          items: { $ref: "#/components/schemas/Annotation" }
          description: Present only with include=annotations.
        metadata:
          description: Unfurl record for the item's link. Present only with include=metadata.
          type: [object, "null"]
...
    Annotation:
      type: object
      required: [kind, value, actor, created_at]
      properties:
        kind: { type: string }
        value: { type: [string, "null"] }
        actor: { type: [string, "null"] }
        created_at: { type: [string, "null"], format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - invalid_parameter
                - invalid_cursor
                - unauthorized
                - not_found
                - method_not_allowed
                - unsupported_media_type
                - payload_too_large
                - internal_error
            message: { type: string }
  responses:
    BadRequest:
      description: A parameter was missing, malformed, or unrecognized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: A token is configured and the request did not present it
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...